Implement Ext in crates
