+++ core-processor/src/lib.rs
@@ -50,7 +50,6 @@ pub const UNAVAILABLE_DEST_EXIT_CODE: ExitCode = 2;
 pub const RE_INIT_EXIT_CODE: ExitCode = 3;
 
 pub use executor::execute_wasm;
-pub use ext::Ext;
 pub use ext::ProcessorExt;
 pub use handler::handle_journal;
 pub use processor::process;
