+++ core/src/costs.rs
@@ -118,10 +118,10 @@ pub struct HostFnWeights {
 #[macro_export]
 macro_rules! charge_gas_token {
     ($ext:expr, $costs:expr) => {{
-        let token = $costs.token(&$ext.host_fn_weights);
+        let token = $costs.token(&$ext.host_fn_weights());
         (
-            $ext.gas_counter.charge_token(token),
-            $ext.gas_allowance_counter.charge_token(token),
+            $ext.gas_counter().charge_token(token),
+            $ext.gas_allowance_counter().charge_token(token),
         )
     }};
 }
