+++ core/src/env.rs
@@ -19,15 +19,17 @@
 //! Environment for running a module.
 
 use crate::{
-    costs::RuntimeCosts,
+    charge_gas_token,
+    costs::{HostFnWeights, RuntimeCosts},
+    gas::{ChargeResult, GasAllowanceCounter, GasCounter, ValueCounter},
     ids::{MessageId, ProgramId},
-    memory::{Memory, WasmPageNumber},
-    message::{ExitCode, HandlePacket, InitPacket, ReplyPacket},
+    memory::{AllocationsContext, Memory, WasmPageNumber},
+    message::{ExitCode, HandlePacket, InitPacket, MessageContext, ReplyPacket},
 };
 use alloc::rc::Rc;
 use codec::{Decode, Encode};
 use core::cell::RefCell;
-use gear_core_errors::CoreError;
+use gear_core_errors::{CoreError, ExtError, TerminationReason};
 
 /// Page access rights.
 #[derive(Clone, Debug, Encode, Decode, PartialEq, Eq, Copy)]
@@ -65,10 +67,26 @@ pub trait Ext {
     fn origin(&mut self) -> Result<ProgramId, Self::Error>;
 
     /// Initialize a new incomplete message for another program and return its handle.
-    fn send_init(&mut self) -> Result<usize, Self::Error>;
+    fn send_init(&mut self) -> Result<usize, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::SendInit)?;
+        let result = self
+            .message_context()
+            .send_init()
+            .map_err(ExtError::Message);
+
+        self.return_and_store_err(result.map(|v| v as usize))
+    }
 
     /// Push an extra buffer into message payload by handle.
-    fn send_push(&mut self, handle: usize, buffer: &[u8]) -> Result<(), Self::Error>;
+    fn send_push(&mut self, handle: usize, buffer: &[u8]) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::SendPush(buffer.len() as u32))?;
+        let result = self
+            .message_context()
+            .send_push(handle as u32, buffer)
+            .map_err(ExtError::Message);
+
+        self.return_and_store_err(result)
+    }
 
     /// Complete message and send it to another program.
     fn send_commit(&mut self, handle: usize, msg: HandlePacket) -> Result<MessageId, Self::Error>;
@@ -80,7 +98,15 @@ pub trait Ext {
     }
 
     /// Push an extra buffer into reply message.
-    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error>;
+    fn reply_push(&mut self, buffer: &[u8]) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Reply(buffer.len() as u32))?;
+        let result = self
+            .message_context()
+            .reply_push(buffer)
+            .map_err(ExtError::Message);
+
+        self.return_and_store_err(result)
+    }
 
     /// Complete reply message and send it to source program.
     fn reply_commit(&mut self, msg: ReplyPacket) -> Result<MessageId, Self::Error>;
@@ -91,18 +117,27 @@ pub trait Ext {
     }
 
     /// Read the message id, if current message is a reply.
-    fn reply_to(&mut self) -> Result<Option<(MessageId, ExitCode)>, Self::Error>;
+    fn reply_to(&mut self) -> Result<Option<(MessageId, ExitCode)>, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::ReplyTo)?;
+        Ok(self.message_context().current().reply())
+    }
 
     /// Get the source of the message currently being handled.
-    fn source(&mut self) -> Result<ProgramId, Self::Error>;
+    fn source(&mut self) -> Result<ProgramId, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Source)?;
+        Ok(self.message_context().current().source())
+    }
 
     /// Terminate the program and transfer all available value to the address.
     fn exit(&mut self, value_destination: ProgramId) -> Result<(), Self::Error>;
 
-    /// Get the id of the message currently being handled.
-    fn message_id(&mut self) -> Result<MessageId, Self::Error>;
+    /// Get the [`MessageId`] of the message currently being handled.
+    fn message_id(&mut self) -> Result<MessageId, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::MsgId)?;
+        Ok(self.message_context().current().id())
+    }
 
-    /// Get the id of program itself
+    /// Get the [`ProgramId`] of program itself
     fn program_id(&mut self) -> Result<ProgramId, Self::Error>;
 
     /// Free specific memory page.
@@ -117,40 +152,123 @@ pub trait Ext {
     fn debug(&mut self, data: &str) -> Result<(), Self::Error>;
 
     /// Interrupt the program, saving it's state.
-    fn leave(&mut self) -> Result<(), Self::Error>;
+    fn leave(&mut self) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Leave)?;
+        Ok(())
+    }
 
     /// Access currently handled message payload.
-    fn msg(&mut self) -> &[u8];
+    fn msg(&mut self) -> &[u8] {
+        self.message_context().current().payload()
+    }
 
     /// Default gas host call.
-    fn gas(&mut self, amount: u32) -> Result<(), Self::Error>;
+    fn gas(&mut self, amount: u32) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::MeteringBlock(amount))
+    }
 
     /// Charge some extra gas.
-    fn charge_gas(&mut self, amount: u32) -> Result<(), Self::Error>;
+    fn charge_gas(&mut self, amount: u32) -> Result<(), Self::Error> {
+        use ChargeResult::*;
+
+        let common_charge = self.gas_counter().charge(amount as u64);
+        let allowance_charge = self.gas_allowance_counter().charge(amount as u64);
+
+        let res = match (common_charge, allowance_charge) {
+            (NotEnough, _) => Err(ExtError::GasLimitExceeded),
+            (Enough, NotEnough) => Err(ExtError::TerminationReason(
+                TerminationReason::GasAllowanceExceeded,
+            )),
+            (Enough, Enough) => Ok(()),
+        };
+
+        self.return_and_store_err(res)
+    }
 
     /// Charge gas by `RuntimeCosts` token.
-    fn charge_gas_runtime(&mut self, costs: RuntimeCosts) -> Result<(), Self::Error>;
+    fn charge_gas_runtime(&mut self, costs: RuntimeCosts) -> Result<(), Self::Error> {
+        use ChargeResult::*;
+        let (common_charge, allowance_charge) = charge_gas_token!(self, costs);
+
+        let res = match (common_charge, allowance_charge) {
+            (NotEnough, _) => Err(ExtError::GasLimitExceeded),
+            (Enough, NotEnough) => Err(ExtError::TerminationReason(
+                TerminationReason::GasAllowanceExceeded,
+            )),
+            (Enough, Enough) => Ok(()),
+        };
+
+        self.return_and_store_err(res)
+    }
 
     /// Refund some gas.
-    fn refund_gas(&mut self, amount: u32) -> Result<(), Self::Error>;
+    fn refund_gas(&mut self, amount: u32) -> Result<(), Self::Error> {
+        if self.gas_counter().refund(amount as u64) == ChargeResult::Enough {
+            self.gas_allowance_counter().refund(amount as u64);
+            Ok(())
+        } else {
+            self.return_and_store_err(Err(ExtError::TooManyGasAdded))
+        }
+    }
 
     /// Tell how much gas is left in running context.
-    fn gas_available(&mut self) -> Result<u64, Self::Error>;
+    fn gas_available(&mut self) -> Result<u64, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::GasAvailable)?;
+        Ok(self.gas_counter().left())
+    }
 
     /// Value associated with message.
-    fn value(&mut self) -> Result<u128, Self::Error>;
+    fn value(&mut self) -> Result<u128, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Value)?;
+        Ok(self.message_context().current().value())
+    }
 
     /// Tell how much value is left in running context.
-    fn value_available(&mut self) -> Result<u128, Self::Error>;
+    fn value_available(&mut self) -> Result<u128, Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::ValueAvailable)?;
+        Ok(self.value_counter().left())
+    }
 
     /// Interrupt the program and reschedule execution.
-    fn wait(&mut self) -> Result<(), Self::Error>;
+    fn wait(&mut self) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Wait)?;
+        Ok(())
+    }
 
     /// Wake the waiting message and move it to the processing queue.
-    fn wake(&mut self, waker_id: MessageId) -> Result<(), Self::Error>;
+    fn wake(&mut self, waker_id: MessageId) -> Result<(), Self::Error> {
+        self.charge_gas_runtime(RuntimeCosts::Wake)?;
+        let result = self
+            .message_context()
+            .wake(waker_id)
+            .map_err(ExtError::Wake);
+
+        self.return_and_store_err(result)
+    }
 
     /// Send init message to create a new program
     fn create_program(&mut self, packet: InitPacket) -> Result<ProgramId, Self::Error>;
+
+    /// Get mut ref of [`GasCounter`]
+    fn gas_counter(&mut self) -> &mut GasCounter;
+
+    /// Get mut ref of [`GasAllowanceCounter`]
+    fn gas_allowance_counter(&mut self) -> &mut GasAllowanceCounter;
+
+    /// Get mut ref of [`ValueCounter`]
+    fn value_counter(&mut self) -> &mut ValueCounter;
+
+    /// Get mut ref of [`MessageContext`]
+    fn message_context(&mut self) -> &mut MessageContext;
+
+    /// Get mut ref of [`AllocationsContext`]
+    fn allocations_context(&mut self) -> &mut AllocationsContext;
+
+    /// Get ref of [`HostFnWeights`]
+    fn host_fn_weights(&self) -> &HostFnWeights;
+
+    /// Return result and store error info in field
+    fn return_and_store_err<T>(&mut self, result: Result<T, ExtError>) -> Result<T, Self::Error>;
 }
 
 /// An error occurred during [`ExtCarrier::with_fallible`] which should be called only when inner is set
@@ -241,7 +359,7 @@ impl<E: Ext> Clone for ClonedExtCarrier<E> {
 mod tests {
     use super::*;
     use core::fmt;
-    use gear_core_errors::TerminationReason;
+    use gear_core_errors::{CoreError, TerminationReason};
 
     #[derive(Debug)]
     struct AllocError;
@@ -266,12 +384,74 @@ mod tests {
     // to call `fn with<R>(&self, f: impl FnOnce(&mut E) -> R) -> R`.
     // For example, returns the field of ext's inner value.
     fn converter(e: &mut ExtImplementedStruct) -> u8 {
-        e.0
+        e.inner
     }
 
     /// Struct with internal value to interact with ExtCarrier
-    #[derive(Debug, PartialEq, Clone, Copy)]
-    struct ExtImplementedStruct(u8);
+    struct ExtImplementedStruct {
+        inner: u8,
+        pub gas_counter: GasCounter,
+        /// Gas allowance counter.
+        pub gas_allowance_counter: GasAllowanceCounter,
+        /// Value counter.
+        pub value_counter: ValueCounter,
+        /// Allocations context.
+        pub allocations_context: AllocationsContext,
+        /// Message context.
+        pub message_context: MessageContext,
+        /// Weights of host functions.
+        pub host_fn_weights: HostFnWeights,
+    }
+
+    impl PartialEq for ExtImplementedStruct {
+        fn eq(&self, other: &Self) -> bool {
+            self.inner == other.inner
+        }
+    }
+
+    impl Clone for ExtImplementedStruct {
+        fn clone(&self) -> Self {
+            Self {
+                inner: self.inner,
+                gas_counter: GasCounter::new(0),
+                gas_allowance_counter: GasAllowanceCounter::new(0),
+                value_counter: ValueCounter::new(0),
+                allocations_context: AllocationsContext::new(
+                    Default::default(),
+                    0.into(),
+                    1.into(),
+                ),
+                message_context: MessageContext::new(Default::default(), Default::default(), None),
+                host_fn_weights: Default::default(),
+            }
+        }
+    }
+
+    impl fmt::Debug for ExtImplementedStruct {
+        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+            f.debug_struct("ExtImplementedStruct")
+                .field("inner", &self.inner)
+                .finish()
+        }
+    }
+
+    impl ExtImplementedStruct {
+        fn new() -> Self {
+            Self {
+                inner: 0,
+                gas_counter: GasCounter::new(0),
+                gas_allowance_counter: GasAllowanceCounter::new(0),
+                value_counter: ValueCounter::new(0),
+                allocations_context: AllocationsContext::new(
+                    Default::default(),
+                    0.into(),
+                    1.into(),
+                ),
+                message_context: MessageContext::new(Default::default(), Default::default(), None),
+                host_fn_weights: Default::default(),
+            }
+        }
+    }
 
     /// Empty Ext implementation for test struct
     impl Ext for ExtImplementedStruct {
@@ -369,14 +549,44 @@ mod tests {
         fn create_program(&mut self, _packet: InitPacket) -> Result<ProgramId, Self::Error> {
             Ok(Default::default())
         }
+        fn gas_counter(&mut self) -> &mut GasCounter {
+            &mut self.gas_counter
+        }
+
+        fn gas_allowance_counter(&mut self) -> &mut GasAllowanceCounter {
+            &mut self.gas_allowance_counter
+        }
+
+        fn value_counter(&mut self) -> &mut ValueCounter {
+            &mut self.value_counter
+        }
+
+        fn message_context(&mut self) -> &mut MessageContext {
+            &mut self.message_context
+        }
+
+        fn allocations_context(&mut self) -> &mut AllocationsContext {
+            &mut self.allocations_context
+        }
+
+        fn host_fn_weights(&self) -> &HostFnWeights {
+            &self.host_fn_weights
+        }
+
+        fn return_and_store_err<T>(
+            &mut self,
+            _result: Result<T, ExtError>,
+        ) -> Result<T, Self::Error> {
+            Err(AllocError)
+        }
     }
 
     #[test]
     fn create_and_unwrap_ext_carrier() {
-        let ext_implementer = ExtImplementedStruct(0);
-        let ext = ExtCarrier::new(ext_implementer);
+        let ext_implementer = ExtImplementedStruct::new();
+        let ext = ExtCarrier::new(ext_implementer.clone());
 
-        assert_eq!(ext.0, Rc::new(RefCell::new(Some(ext_implementer))));
+        assert_eq!(ext.0, Rc::new(RefCell::new(Some(ext_implementer.clone()))));
 
         let inner = ext.into_inner();
 
@@ -385,7 +595,7 @@ mod tests {
 
     #[test]
     fn calling_fn_within_inner_ext() {
-        let ext_implementer = ExtImplementedStruct(0);
+        let ext_implementer = ExtImplementedStruct::new();
         let ext = ExtCarrier::new(ext_implementer);
         let ext_clone = ext.cloned();
 
@@ -395,7 +605,7 @@ mod tests {
 
     #[test]
     fn calling_fn_when_ext_unwrapped() {
-        let ext = ExtCarrier::new(ExtImplementedStruct(0));
+        let ext = ExtCarrier::new(ExtImplementedStruct::new());
         let ext_clone = ext.cloned();
 
         let _ = ext.into_inner();
@@ -404,7 +614,7 @@ mod tests {
 
     #[test]
     fn calling_fn_when_dropped_ext() {
-        let ext = ExtCarrier::new(ExtImplementedStruct(0));
+        let ext = ExtCarrier::new(ExtImplementedStruct::new());
         let ext_clone = ext.cloned();
 
         drop(ext);
@@ -416,8 +626,8 @@ mod tests {
     #[allow(clippy::redundant_clone)]
     /// Test that ext's clone still refers to the same inner object as the original one
     fn ext_cloning() {
-        let ext_implementer = ExtImplementedStruct(0);
-        let ext = ExtCarrier::new(ext_implementer);
+        let ext_implementer = ExtImplementedStruct::new();
+        let ext = ExtCarrier::new(ext_implementer.clone());
         let ext_clone = ext.cloned();
 
         assert_eq!(ext_clone.0 .0, Rc::new(RefCell::new(Some(ext_implementer))));
@@ -425,8 +635,8 @@ mod tests {
 
     #[test]
     fn unwrap_ext_with_dropped_clones() {
-        let ext_implementer = ExtImplementedStruct(0);
-        let ext = ExtCarrier::new(ext_implementer);
+        let ext_implementer = ExtImplementedStruct::new();
+        let ext = ExtCarrier::new(ext_implementer.clone());
         let ext_clone1 = ext.cloned();
         let ext_clone2 = ext_clone1.clone();
