+++ gear-test/src/main.rs
@@ -18,13 +18,14 @@
 
 mod address;
 mod check;
+mod ext;
 mod js;
 mod manager;
 mod proc;
 mod sample;
 
 use clap::Parser;
-use core_processor::Ext;
+use ext::Ext;
 use gear_backend_wasmtime::WasmtimeEnvironment;
 use manager::InMemoryExtManager;
