+++ pallets/gear/src/mock.rs
@@ -17,13 +17,12 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate as pallet_gear;
-use crate::{ext::LazyPagesExt, manager::ExtManager};
+use crate::{ext::Ext, manager::ExtManager};
 use common::lazy_pages;
 use common::Origin as _;
 use core_processor::{
     common::{DispatchOutcome, JournalNote},
     configs::BlockInfo,
-    Ext,
 };
 use frame_support::traits::{Currency, FindAuthor, OnFinalize, OnIdle, OnInitialize};
 use frame_support::{construct_runtime, parameter_types};
@@ -258,31 +257,17 @@ pub fn calc_handle_gas_spent(source: H256, dest: H256, payload: Vec<u8>) -> (u64
     let existential_deposit =
         <Test as pallet_gear::Config>::Currency::minimum_balance().unique_saturated_into();
 
-    let journal = if lazy_pages_enabled {
-        core_processor::process::<LazyPagesExt, SandboxEnvironment<_>>(
-            Some(actor),
-            dispatch.into_stored().into_incoming(initial_gas),
-            block_info,
-            existential_deposit,
-            ProgramId::from_origin(source),
-            ProgramId::from_origin(dest),
-            u64::MAX,
-            <Test as pallet_gear::Config>::OutgoingLimit::get(),
-            schedule.host_fn_weights.into_core(),
-        )
-    } else {
-        core_processor::process::<Ext, SandboxEnvironment<_>>(
-            Some(actor),
-            dispatch.into_stored().into_incoming(initial_gas),
-            block_info,
-            existential_deposit,
-            ProgramId::from_origin(source),
-            ProgramId::from_origin(dest),
-            u64::MAX,
-            <Test as pallet_gear::Config>::OutgoingLimit::get(),
-            schedule.host_fn_weights.into_core(),
-        )
-    };
+    let journal = core_processor::process::<Ext, SandboxEnvironment<_>>(
+        Some(actor),
+        dispatch.into_stored().into_incoming(initial_gas),
+        block_info,
+        existential_deposit,
+        ProgramId::from_origin(source),
+        ProgramId::from_origin(dest),
+        u64::MAX,
+        <Test as pallet_gear::Config>::OutgoingLimit::get(),
+        schedule.host_fn_weights.into_core(),
+    );
 
     let mut gas_burned: u64 = 0;
     let mut gas_to_send: u64 = 0;
